+++ src/js/types.ts
@@ -549,7 +549,7 @@ export interface API {
     buffer: Uint8Array,
     filename: string,
     installDir: string,
-    metadata?: ReadonlyMap<string, string>,
+    metadata?: Record<string, string> | PyProxy,
   ) => Promise<void>;
   _Comlink: any;
 
@@ -596,10 +596,11 @@ export type PackageManagerAPI = Pick<
   | "sitepackages"
   | "defaultLdLibraryPath"
   | "version"
+  | "pyVersionTuple"
 > & {
   config: Pick<
     PyodideConfigWithDefaults,
-    "packageCacheDir" | "packageBaseUrl" | "cdnUrl"
+    "packageCacheDir" | "packageBaseUrl" | "cdnUrl" | "preloadSharedLibraries"
   >;
 };
 /**
@@ -607,6 +608,7 @@ export type PackageManagerAPI = Pick<
  */
 export type PackageManagerModule = Pick<
   PyodideModule,
+  | "FS"
   | "PATH"
   | "LDSO"
   | "stringToNewUTF8"